Refactor SharedViewModel
