+++ app/src/main/java/ch/heigvd/pro/b04/android/Poll/PollViewModel.java
@@ -3,9 +3,23 @@ package ch.heigvd.pro.b04.android.Poll;
 import android.app.Application;
 
 import androidx.annotation.NonNull;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+
+import ch.heigvd.pro.b04.android.Datamodel.Question;
 
 public class PollViewModel extends SharedViewModel {
+    private MutableLiveData<Question> questionToView = new MutableLiveData<>();
+
     public PollViewModel(@NonNull Application application, String token, int idModerator, int idPoll) {
         super(application, token, idModerator, idPoll);
+    }
+
+    public void goToQuestion(Question question) {
+        questionToView.postValue(question);
+    }
+
+    public LiveData<Question> getQuestionToView() {
+        return questionToView;
     }
 }
